Исправление багов\nИзменил последовательность открытия и закрытия транзакций. Убрал конкатенацию строк в hql запросах и привел их к стандарту

+++ src/main/java/com/kindcat/archivemedo/beans/ImplBeans.java
@@ -9,9 +9,9 @@ public interface ImplBeans {
 
 
      */
-    public int getBeanIdOrg();
+    public Short getBeanIdOrg();
 
-    public void setBeanIdOrg(int idOrg);
+    public void setBeanIdOrg(Short idOrg);
 
     public String getBeanNameOrg();
