+++ src/main/java/com/kindcat/archivemedo/beans/SuperBeans.java
@@ -67,12 +67,12 @@ public class SuperBeans implements ImplBeans {
      */
 
     @Override
-    public int getBeanIdOrg() {
+    public Short getBeanIdOrg() {
         return memberBean.getBeansIdOrg();
     }
 
     @Override
-    public void setBeanIdOrg(int idOrg) {
+    public void setBeanIdOrg(Short idOrg) {
         memberBean.setBeansIdOrg(idOrg);
     }
